applied pvgrub patch on xen-tools-3.3.1
git-svn-id: https://subversion.fem.tu-ilmenau.de/repository/fem-overlay/trunk@48 d0c66219-219a-4382-8d04-2ce9737ae069

+++ app-emulation/xen-tools/xen-tools-3.3.1-r1.ebuild
@@ -7,14 +7,19 @@ inherit flag-o-matic eutils multilib python
 
 DESCRIPTION="Xend daemon and tools"
 HOMEPAGE="http://xen.org/"
-SRC_URI="http://bits.xensource.com/oss-xen/release/${PV}/xen-${PV}.tar.gz"
+SRC_URI="http://bits.xensource.com/oss-xen/release/${PV}/xen-${PV}.tar.gz
+        pvgrub? ( http://alpha.gnu.org/gnu/grub/grub-0.97.tar.gz
+		http://www.zlib.net/zlib-1.2.3.tar.gz
+		http://www.kernel.org/pub/software/utils/pciutils/pciutils-2.2.9.tar.bz2
+		http://download.savannah.gnu.org/releases/lwip/lwip-1.3.0.tar.gz
+		ftp://sources.redhat.com/pub/newlib/newlib-1.16.0.tar.gz )"
 #	vtpm? ( mirror://berlios/tpm-emulator/${TPMEMUFILE} )"
 S="${WORKDIR}/xen-${PV}"
 
 LICENSE="GPL-2"
 SLOT="0"
 KEYWORDS="~amd64 ~x86"
-IUSE="doc debug screen custom-cflags pygrub hvm api acm flask"
+IUSE="doc debug screen custom-cflags pygrub pvgrub hvm api acm flask"
 
 CDEPEND="dev-lang/python
 	sys-libs/zlib
@@ -121,6 +126,15 @@ src_unpack() {
 			-i {} \;
 	fi
 
+	if use pvgrub; then
+		sed -i \
+		-e 's/WGET=.*/WGET=cp -t . /' \
+		-e "s;\$(XEN_EXTFILES_URL);${DISTDIR};" \
+		-e 's/$(LD)/$(LD) LDFLAGS=/' \
+		-e 's;install-grub: pv-grub;install-grub:;' \
+		stubdom/Makefile
+	fi
+
 	# Disable hvm support on systems that don't support x86_32 binaries.
 	if ! use hvm; then
 		chmod 644 tools/check/check_x11_devel
@@ -158,6 +172,13 @@ src_compile() {
 
 	emake -C tools ${myopt} || die "compile failed"
 
+	if use pvgrub; then
+		emake -C stubdom pv-grub || die "compile pv-grub_${XEN_TARGET_ARCH} failed"
+		if use amd64; then
+			emake XEN_TARGET_ARCH="x86_32" -C stubdom pvgrub || die "compile pv-grub_x86_32 failed"
+		fi
+	fi
+
 	if use doc; then
 		sh ./docs/check_pkgs || die "package check failed"
 		emake docs || die "compiling docs failed"
@@ -170,6 +191,13 @@ src_compile() {
 src_install() {
 	make DESTDIR="${D}" DOCDIR="/usr/share/doc/${PF}" XEN_PYTHON_NATIVE_INSTALL=y install-tools  \
 		|| die "install failed"
+
+	if use pvgrub; then
+		emake DESTDIR="${D}" -C stubdom install-grub || die "install pvgrub_${XEN_TARGET_ARCH} failed"
+		if use amd64; then
+			emake XEN_TARGET_ARCH="x86_32" -C stubdom install-grub || die "install pv-grub_x86_32 failed"
+		fi
+	fi
 
 	# Remove RedHat-specific stuff
 	rm -rf "${D}"/etc/sysconfig
